final code

+++ server/Controllers/Section.js
@@ -48,7 +48,7 @@ exports.createSection = async (req, res) => {
 exports.updateSection = async (req, res) => {
     try {
         //data input
-        const { sectionName, sectionId } = req.body;
+        const { sectionName, sectionId,courseId  } = req.body;
         //data validation
         if (!sectionName || !sectionId) {
             return res.status(400).json({
@@ -59,12 +59,21 @@ exports.updateSection = async (req, res) => {
         //update data
         
         const section = await Section.findByIdAndUpdate( sectionId , { sectionName }, { new: true });
+        
+        const course = await Course.findById(courseId)
+		.populate({
+			path:"courseContent",
+			populate:{
+				path:"subSection",
+			},
+		})
+		.exec();
         //return res
        
         return res.status(200).json({
             success: true,
             message: 'Section Updated Successfully',
-            section
+            data:course,
         });
 
 
@@ -79,22 +88,44 @@ exports.updateSection = async (req, res) => {
 exports.deleteSection = async (req, res) => {
     try {
         //get ID - assuming that we are sending ID in params
-        const { sectionId } = req.params;
-        console.log(sectionId);
-         //use findByIdandDelete
-        const deletedSection= await Section.findByIdAndDelete(sectionId);
+        const { sectionId,courseId  } = req.body;
+        // console.log(sectionId);
+        
+        await Course.findByIdAndUpdate(courseId, {
+			$pull: {
+				courseContent: sectionId,
+			}
+		})
+
+        const section = await Section.findById(sectionId);
+		console.log(sectionId, courseId);
+		if(!section) {
+			return res.status(404).json({
+				success:false,
+				message:"Section not Found",
+			})
+		}
+        //delete sub section
+        if (section.subSection.length > 0) {
+            await SubSection.deleteMany({ _id: { $in: section.subSection } });
+          }
+          
+        await Section.findByIdAndDelete(sectionId);
+        //find the updated course and return 
+		const course = await Course.findById(courseId).populate({
+			path:"courseContent",
+			populate: {
+				path: "subSection"
+			}
+		})
+		.exec();
          //TODO[Testing]: do we need to delete the entry from the course schema ??
          //return response
-         if(!deletedSection){
-            return res.status(500).json({
-                success:false,
-                message:"Unable to delete Section, please try again",
-               
-            });
-         }
+        
          return res.status(200).json({
              success:true,
              message:"Section Deleted Successfully",
+             data:course
          })
